% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.r
\name{dijkstra}
\alias{dijkstra}
\title{calculate the shortest path}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{a data.frame}

\item{init_node}{a number}
}
\value{
the shortest path from the init_node to every other node
}
\description{
calculate the shortest path
}
\examples{
dijkstra(wiki_graph,1)
dijkstra(wiki_graph,3)
}
\seealso{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
