% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.r
\name{euclidean}
\alias{euclidean}
\title{calculate the greatest common divisor of two numbers}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{a number}

\item{b}{a number}
}
\value{
the greatest common divisor of two numbers
}
\description{
calculate the greatest common divisor of two numbers
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
\seealso{
\url{https://en.wikipedia.org/wiki/Euclidean_algrithm}
}
