\name{euclir_1-package}
\alias{euclir_1-package}
\alias{euclir_1}
\docType{package}
\title{
\packageTitle{euclir_1}
}
\description{
\packageDescription{euclir_1}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{euclir_1}
\packageIndices{euclir_1}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{euclir_1}

Maintainer: \packageMaintainer{euclir_1}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
