% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Graph including nodes and weight of each edge}
\format{A data frame with 18 rows and 3 variables:
   \describe{
      \item{v1} {the beginning node in a edge}
      \item{v2} {the end node in a edge}
      \item{w}  {the weight of the edge from v1 to v2}

   }
   @source \url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}}
\usage{
wiki_graph
}
\description{
A dataset containing the nodes and edge weight of a graph
}
\keyword{datasets}
